requirements.txt
torch>=2.0
transformers==4.36
accelerate>=0.25
opencv-python==4.8
albumentations>=1.3
scikit-learn>=1.3
tqdm>=4.66
numpy==1.26.4
Pillow>=10.0
requests>=2.31
streamlit
